Switch to 3.11 and skip existing files in dataset formatting

+++ format_dataset.py
@@ -11,11 +11,11 @@ FORMATTED_DATASET_PATH = Path("/storage/tpenner/french_pronunciation_dataset/")
 if not ORIGINAL_DATASET_PATH.exists():
     raise Exception("Original dataset path doesn't exist")
 
-FORMATTED_DATASET_PATH.mkdir()
+FORMATTED_DATASET_PATH.mkdir(exist_ok=True)
 for split in SPLITS:
     logger.info(f"Processing {split} split")
 
-    (FORMATTED_DATASET_PATH / split).mkdir()
+    (FORMATTED_DATASET_PATH / split).mkdir(exist_ok=True)
 
     split_path = ORIGINAL_DATASET_PATH / f"{split}.tsv"
     df = pd.read_table(split_path)
@@ -31,9 +31,18 @@ for split in SPLITS:
         new_audio_path = speaker_dir / file_name.split("_")[-1]
         new_text_path = new_audio_path.with_suffix(".lab")
 
-        # Hard link from original audio to path in dataset
-        new_audio_path.hardlink_to(original_audio_path)
+        if new_audio_path.exists() and new_text_path.exists():
+            continue
+        elif new_audio_path.exists() and not new_text_path.exists():
+            logger.warning(f"Missing text file for {new_audio_path}")
+        elif not new_audio_path.exists() and new_text_path.exists():
+            logger.warning(f"Missing audio file for {new_text_path}")
 
-        with open(new_text_path, "w") as f:
-            # Common voice recomends removing smart apostrophes
-            f.write(sentence_text.replace("‘", "'").replace("’", "'").strip())
+        if not new_audio_path.exists():
+            # Hard link from original audio to path in dataset
+            new_audio_path.hardlink_to(original_audio_path)
+
+        if not new_text_path.exists():
+            with open(new_text_path, "w") as f:
+                # Common voice recomends removing smart apostrophes
+                f.write(sentence_text.replace("‘", "'").replace("’", "'").strip())
